modificando un par de fors
