--- src/middlewares/cloudinary/cloudinary.middleware.ts
@@ -1,4 +1,0 @@
-// import * as multer from 'multer';
-
-// // Configurar Multer para almacenar los archivos en memoria
-// export const upload = multer({ storage: multer.memoryStorage() });
