+++ src/modules/events/dto/update-event.dto.ts
@@ -1,4 +1,3 @@
-import { PartialType } from '@nestjs/mapped-types';
 import { IsOptional, IsString, IsNotEmpty, IsArray, IsNumber, Matches, IsDateString, IsDate } from 'class-validator';
 import { TypeEvents } from 'src/utils/enum';
 
@@ -23,9 +22,6 @@ export class UpdateEventDto {
   @Matches(/^(0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]$/, { message: 'El formato de hora debe ser HH:MM (24 horas)' })
   time?: string;
 
-  // @IsOptional()
-  // location?: { lat: string; log: string };
-
   @IsNumber()
   @IsNotEmpty()
   lat: number
@@ -46,10 +42,6 @@ export class UpdateEventDto {
   @IsOptional()
   @IsNumber()
   amount?: number;
-
-  // @IsOptional()
-  // @IsString()
-  // comment?: string;
 
   @IsNotEmpty()
   @IsDate()
