+++ src/modules/events/events.service.ts
@@ -2,7 +2,8 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaService } from '../../prisma.service';
 import { CreateEventDto } from 'src/modules/events/dto/create-event.dto';
 import { UpdateEventDto } from 'src/modules/events/dto/update-event.dto';
-import { deleteImgCloudinary, filterEventsRadius, uploadFilesToCloudinary } from 'src/utils/utils';
+import {filterEventsRadius} from 'src/utils/utils';
+import { deleteImgCloudinary, uploadFilesToCloudinary } from 'src/utils/utils.cloudinary';
 import { QueryEventsDto } from './dto/query-event.dto';
