--- src/modules/events/pipes/validateuser/validateuser.pipe.spec.ts
@@ -1,7 +1,0 @@
-import { ValidateuserPipe } from './validateuser.pipe';
-
-describe('ValidateuserPipe', () => {
-  it('should be defined', () => {
-    expect(new ValidateuserPipe()).toBeDefined();
-  });
-});
