--- src/modules/events/pipes/validateuser/validateuser.pipe.ts
@@ -1,16 +1,0 @@
-import { ArgumentMetadata, HttpException, HttpStatus, Injectable, PipeTransform } from '@nestjs/common';
-
-@Injectable()
-export class ValidateuserPipe implements PipeTransform {
-  transform(value: any, metadata: ArgumentMetadata) {
-
-
-    const num = parseInt(value.age.toString(),10);
-
-    if (isNaN(num)) {
-      throw new HttpException('el valor debe ser un numero',HttpStatus.BAD_REQUEST);
-    }
-
-    return {...value, age : num};
-  }
-}
